# docs/swagger.yaml
definitions:
  delete.Response:
    properties:
      error:
        items:
          type: string
        type: array
    type: object
  redirect.Response:
    properties:
      error:
        items:
          type: string
        type: array
    type: object
  save.Request:
    properties:
      alias:
        type: string
      url:
        type: string
    required:
    - url
    type: object
  save.Response:
    properties:
      alias:
        type: string
      error:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  description: REST API to short long urls
  title: URL Shortener API
  version: "1.0"
paths:
  /{alias}:
    get:
      consumes:
      - application/json
      description: Redirect to URL by shorted alias
      parameters:
      - description: Short URL alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Temporary Redirect
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/redirect.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/redirect.Response'
      tags:
      - redirect
  /url:
    post:
      consumes:
      - application/json
      description: Converts long URL to short alias
      parameters:
      - description: URL data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/save.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/save.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/save.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/save.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/save.Response'
      tags:
      - url
  /url/{alias}:
    delete:
      consumes:
      - application/json
      description: Delete URL with shorted alias
      parameters:
      - description: Short URL alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delete.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/delete.Response'
      tags:
      - url
swagger: "2.0"